requirements.txt
# requirements.txt

pandas
numpy
scikit-learn
scipy
scikit-surprise
fastapi
uvicorn
joblib
python-dotenv
asyncpg
sqlalchemy[asyncio]
python-jose[cryptography]
passlib[bcrypt]
python-multipart